# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.CreateReq:
    properties:
      color:
        type: string
      name:
        maxLength: 128
        minLength: 1
        type: string
    required:
    - name
    type: object
...
  handler.ErrorResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
...
  handler.ProjectCreateData:
    properties:
      project:
        $ref: '#/definitions/models.Project'
    type: object
  handler.ProjectCreateResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/handler.ProjectCreateData'
      msg:
        type: string
    type: object
  handler.ProjectDeleteData:
    properties:
      id:
        type: integer
      proj_affected:
        type: integer
      task_affected:
        type: integer
    type: object
  handler.ProjectDeleteResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/handler.ProjectDeleteData'
      msg:
        type: string
    type: object
  handler.ProjectDetailData:
    properties:
      project:
        $ref: '#/definitions/service.ProjectProfile'
    type: object
  handler.ProjectDetailResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/handler.ProjectDetailData'
      msg:
        type: string
    type: object
  handler.ProjectListData:
    properties:
      list:
        items:
          $ref: '#/definitions/service.ProjectSummary'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  handler.ProjectListResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/handler.ProjectListData'
      msg:
        type: string
    type: object
  handler.ProjectUpdateData:
    properties:
      project:
        $ref: '#/definitions/models.Project'
    type: object
  handler.ProjectUpdateResponse:
    properties:
      code:
        type: integer
      count:
        type: integer
      data:
        $ref: '#/definitions/handler.ProjectUpdateData'
      msg:
        type: string
    type: object
...
  handler.UpdateReq:
    properties:
      color:
        maxLength: 16
        type: string
      name:
        maxLength: 128
        minLength: 1
        type: string
      sort_order:
        type: integer
    type: object
...
  models.Project:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  service.ProjectProfile:
    properties:
      color:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      sort_order:
        type: integer
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  service.ProjectSummary:
    properties:
      color:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
  description: 管理API
  title: ToDoList API
  version: "1.0"
paths:
...
  /projects:
    get:
      consumes:
      - application/json
      description: 获取当前用户的项目列表，支持按名称模糊搜索和分页
      parameters:
      - description: 项目名称关键字（模糊搜索）
        in: query
        name: name
        type: string
      - description: 页码（默认1）
        in: query
        name: page
        type: integer
      - description: 每页数量（默认20，最大100）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回项目列表
          schema:
            $ref: '#/definitions/handler.ProjectListResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 搜索项目列表
    post:
      consumes:
      - application/json
      description: 创建新项目，项目名称必需且不能为空
      parameters:
      - description: 项目创建请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功，返回项目信息
          schema:
            $ref: '#/definitions/handler.ProjectCreateResponse'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: 项目重复
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 创建项目
  /projects/{id}:
    delete:
      consumes:
      - application/json
      description: 删除项目及其关联的所有任务
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功，返回删除的项目ID和受影响的行数
          schema:
            $ref: '#/definitions/handler.ProjectDeleteResponse'
        "400":
          description: 非法的项目ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 删除项目
    get:
      consumes:
      - application/json
      description: 根据项目ID获取项目的详细信息，包括名称、颜色和创建时间
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功，返回项目信息
          schema:
            $ref: '#/definitions/handler.ProjectDetailResponse'
        "400":
          description: 非法的项目ID
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 获取项目详情
    patch:
      consumes:
      - application/json
      description: 更新项目的名称、颜色和排序顺序（可选项）
      parameters:
      - description: 项目ID
        in: path
        name: id
        required: true
        type: integer
      - description: 项目更新请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功，返回更新后的项目信息
          schema:
            $ref: '#/definitions/handler.ProjectUpdateResponse'
        "400":
          description: 非法的项目ID或参数格式错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: 未授权或token无效
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: 项目不存在
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: 项目重复
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: 系统错误
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - Bearer: []
      summary: 更新项目信息
...
securityDefinitions:
  Bearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"